==> code/1_Create_List_All_Solvents.ipynb <==
# Make lists of all Solvents and Catalysts (+Reagents)
This jupyter notebook creates lists of e.g. used solvents within the found reactions

## Libraries

from os import listdir
from os.path import isfile, join
import pandas as pd
import collections

## Solvents

#open all compound reaction files
onlyfiles = [f for f in listdir('../data/raw/compounds/') if isfile(join('../data/raw/compounds/', f)) if 'DS_' not in f]

#go through all files
all_solvent = []
for file in onlyfiles:
    df = pd.read_excel('../data/raw/compounds/'+file)
    all_solvent.extend( df['Solvent (Reaction Details)'].unique())

#save all unique used solvents within all found reactions
all_solvent = list(set(all_solvent))
all_solvent = [x for x in all_solvent if str(x) != 'nan']

print ('Number of solvents: %d' %len(all_solvent))

#Output all used solvents
fp_out = open('../results/1_Solvent_Catalyst_Lists/List_of_all_solvents.tsv','w')
for solvent in all_solvent:
    print (solvent)
    fp_out.write(solvent+'\n')
fp_out.close()

## Catalysts + Reagents

#open all compound reaction files
onlyfiles = [f for f in listdir('../data/raw/compounds/') if isfile(join('../data/raw/compounds/', f)) if 'DS_' not in f]

#lists that contains all reagents/catalysts
all_reagents_and_catalysts = []

#go through all files
for file in onlyfiles:
    print (file)
    #open the file as pandas dataframe
    df = pd.read_excel('../data/raw/compounds/'+file)
    
    #get unique reagent and catalyst
    reagents =  df['Reagent'].unique()
    catalysts =  df['Catalyst'].unique()

    #add to list
    all_reagents_and_catalysts.extend(reagents)
    all_reagents_and_catalysts.extend(catalysts)
    
#keep only unqie
all_reagents_and_catalysts = list(set(all_reagents_and_catalysts))
all_reagents_and_catalysts = [x for x in all_reagents_and_catalysts if str(x) != 'nan']

print ('Number of reagents and catalysts: %d' %len(all_reagents_and_catalysts))

#save list of reagents/catalysts
fp_out = open('../results/1_Solvent_Catalyst_Lists/List_of_all_reagents_and_catalysts.tsv','w')
for r_c in all_reagents_and_catalysts:
    print (r_c)
    fp_out.write(r_c+'\n')
fp_out.close()

## Other conditions

#open all compound reaction files
onlyfiles = [f for f in listdir('../data/raw/compounds/') if isfile(join('../data/raw/compounds/', f)) if 'DS_' not in f]

#lists that contains all reagents/catalysts
all_conditions = []

#go through all files
for file in onlyfiles:
    #file = 'Reaxys_Compound_1.xlsx'
    print (file)
    df = pd.read_excel('../data/raw/compounds/'+file)
    
    #get unique reagent and catalyst
    conditions =  df['Other Conditions'].unique()

    #add to list
    all_conditions.extend(conditions)

    
Count = collections.Counter(all_conditions)

#check categories and output those that are at least >5 times used
keep_categories =[]
for c in Count.most_common(50):
    if str(c[0]) != 'nan' and ';' not in str(c[0]):
        
        if c[1] >=5:
            keep_categories.append(c[0])

print (keep_categories)
print ('Number of conditions: %d' %len(keep_categories))

#save list of conditions
fp_out = open('../results/1_Solvent_Catalyst_Lists/List_of_interesting_conditions.tsv','w')
for r_c in keep_categories:
    print (r_c)
    fp_out.write(r_c+'\n')
fp_out.close()
