==> code/2_Prepare_Input_Files.ipynb <==
# Prepare all lists for later use
- Make a list with all the solvents and their hazardenous
- Make a list of all GHS/H/P codes
- Compile a list of H_code to severity

## Libraries

import pandas as pd
import numpy as np
from os import listdir
from os.path import isfile, join

## Functions and definitions

# No standard functions needed

## Main part

### Solvent file
Make a easy readable list showing the toxicity for each solvent

#read all solvents and their toxicities according to Bryne et al.
#this list was compiled using the previously created list in 1_Create_List_All_Solvents.ipynb
solvents = pd.read_excel('../data/raw/List_of_all_solvents_MCcompiled_MMUannot_Aug2019.xlsx')

# remove last row (just comment)
solvents = solvents.drop(25)

#read all solvents
all_solvents = (solvents.values[:,0])
#read all toxicities (values between 1 and 6)
all_values = (solvents.values[:,1])

#This dictionary contains the toxicity for each solvent (according to Miriam)
solvent_to_value = {}

#go through all solvent names and toxicites and assign them correctly (for mixtures for now assign individual values)
for solv, val in zip(all_solvents,all_values):
    solvs = solv.split(';')

    vals = str(val).split(';')
    for s,v in zip(solvs,vals):

        solvent_to_value[s] = int(v)
print (solvent_to_value)

#Make output list
fp_out = open('../data/processed/List_of_Solvents.tsv','w')
fp_out.write('Solvent\thazardousness\n')
for key in solvent_to_value:
    fp_out.write(key+'\t'+str(solvent_to_value[key])+'\n')
fp_out.close()

## Make list of H_code to toxicity

Hazard codes have either categories (mostly), divisions or other 
see https://www.era-environmental.com/blog/ghs-hazard-classification

- Categories easiest with 1 being the most severe, 4 the least
- Division similar easy
- Other such as "Gases under pressure" got a middle severity since the Signal word is not danger typiacally corresponding to category 3 and 4
- Quite rare category 5 exists (special case --> weigthed 0.1, i.e., lower than 1/4

category_to_severity = {
    
    #Physical hazzards
    'Explosives' : {'Unstable Explosive':6/6, 'Div 1.1':6/6, 'Div 1.2':5/6, 'Div 1.3':4/6, 'Div 1.4':3/6, 'Div 1.5':2./6, 'Div 1.6':1./6},
    'Desensitized explosives' : {'Category 1':4/4, 'Category 2':3/4,'Category 3':2/4, 'Category 4':1/4},
    'Flammable gases': {'1A: Flammable gas, Pyrophoric gas, Chemically unstable gas A,B':1.0, '1B':1.0,'1A, Pyrophoric gas':1.0, '1A, Chemically unstable gas A':0.5, '1A, Chemically unstable gas B':0.5, 'Category 2':3/4},
    'Aerosols': {'Category 1':4/4, 'Category 2':3/4, 'Category 3':2/4},
    'Flammable liquids': {'Category 1':4/4, 'Category 2':3/4,'Category 3':2/4, 'Category 4':1/4},
    'Flammable solids' : {'Category 1':4/4, 'Category 2':3/4},
    'Self-reactive substances and mixtures; Organic peroxides' : {'Type A':5/5, 'Type B':4/5,'Type C, D':3/5, 'Type E, F':2/5, 'Type G':1/5},
    'Pyrophoric liquids; Pyrophoric solids' : {'Category 1':4/4},
    'Self-heating substances and mixtures' : {'Category 1':4/4, 'Category 2':3/4},
    'Substances and mixtures which in contact with water, emit flammable gases': {'Category 1':4/4, 'Category 2':3/4,'Category 3':2/4},
    'Oxidizing gases': {'Category 1':4/4},
    'Oxidizing liquids; Oxidizing solids' : {'Category 1':4/4, 'Category 2':3/4,'Category 3':2/4},
    'Gases under pressure': {'Compressed gas, Liquefied gas, Dissolved gas':0.5, 'Refrigerated liquefied gas':0.5 },
    'Chemicals under pressure' : {'Category 1':4/4, 'Category 2':3/4, 'Category 3':2/4},
    'Corrosive to Metals' : {'Category 1':4/4},
    
    
    #Health hazzards
    'Acute toxicity, oral' : {'Category 1, 2':4/4, 'Category 3':2/4,'Category 4':1/4, 'Category 5':0.1}, ##Contains Category 5
    'Aspiration hazard' : {'Category 1':4/4, 'Category 2':3/4},
    'Acute toxicity, dermal' : {'Category 1, 2':4/4, 'Category 3':2/4,'Category 4':1/4, 'Category 5':0.1}, #Contains Category 5
    'Skin corrosion/irritation' : {'Category 1A, 1B, 1C':4/4, 'Category 2':3/4, 'Category 3':2/4},
    'Sensitization, Skin' : {'Category 1, 1A, 1B' :4/4},
    'Serious eye damage/eye irritation' : {'Category 1':4/4, 'Category 2A':3/4, 'Category 2B':2/4},
    'Acute toxicity, inhalation': {'Category 1, 2':4/4, 'Category 3':2/4,'Category 4':1/4, 'Category 5':0.1}, #Contains Category 5
    'Sensitization, respiratory': {'Category 1, 1A, 1B':4/4},
    'Specific target organ toxicity, single exposure; Respiratory tract irritation' : {'Category 3':2/4},
    'Specific target organ toxicity, single exposure; Narcotic effects': {'Category 3':2/4},
    'Germ cell mutagenicity': {'Category 1A, 1B':4/4, 'Category 2':3/4},
    'Carcinogenicity': {'Category 1A, 1B':4/4, 'Category 2':3/4},
    'Reproductive toxicity': {'Category 1A, 1B':4/4, 'Category 2':3/4},
    'Reproductive toxicity, effects on or via lactation' : { 'Additional category': 1/4},
    'Specific target organ toxicity, single exposure': {'Category 1':4/4, 'Category 2':3/4},
    'Specific target organ toxicity, repeated exposure' : {'Category 1':4/4, 'Category 2':3/4},
    
    #Environment hazzards
    'Hazardous to the aquatic environment, acute hazard' : {'Category 1':4/4, 'Category 2':3/4, 'Category 3':2/4},
    'Hazardous to the aquatic environment, long-term hazard' : {'Category 1':4/4, 'Category 2':3/4, 'Category 3':2/4, 'Category 4':1/4},
    'Hazardous to the ozone layer': {'Category 1':4/4},


    #Combined H_codes
    'H300+H310': 4/4,       #Category 1
    'H300+H330': 4/4,       #Category 1
    'H310+H330': 4/4,       #Category 1
    'H300+H310+H330': 4/4,  #Category 1
    
    'H301+H311': 2/4,       #Category 3
    'H301+H331': 2/4,       #Category 3
    'H311+H331': 2/4,       #Category 3
    'H301+H311+H331': 2/4,  #Category 3
    
    'H302+H312': 1/4,       #Category 4
    'H302+H332': 1/4,       #Category 4
    'H312+H332': 1/4,       #Category 4
    'H302+H312+H332': 1/4,  #Category 4
    
    'H303+H313': 0.1,        #Category 5
    'H303+H333': 0.1,        #Category 5
    'H313+H333': 0.1,        #Category 5
    'H303+H313+H333': 0.1,   #Category 5
    
    'H315+H320': 3/4         #Category 2, 2B	
    
}
    

#read the .xslx file with the corresponding categories/divisions etc.
H_code_list = pd.read_excel('../data/raw/Hcodes_withCategories.xlsx')
H_code_list.head()

#final dictionary that contains the H-code severity
H_code_to_severity = {}

#go through all H-codes
for row in H_code_list.iterrows():
    
    #get hcode
    hcode = row[1][0]
    
    #if 'empty' row (e.g. group header for a class of hazard codes)
    if str(hcode) == 'nan':
        continue

    #get other values
    hazard_class = row[1][2]
    category = row[1][3]
    signal_word = row[1][4]

    #some hcodes are joints of several hcodes; those joint H-codes got exact severity classes
    if '+' not in hcode:
        H_code_to_severity[hcode] = category_to_severity[hazard_class][category]
    else:
        H_code_to_severity[hcode] = category_to_severity[hcode]
    

#create a file with a severity for each H-code
fp_out = open('../data/processed/HCode_Severity.csv','w')
fp_out.write('Hcode,Severity\n')
for code in H_code_to_severity:
    fp_out.write(code+','+str(H_code_to_severity[code])+'\n')
fp_out.close()

### Catalyst File

#open the manual annotated catalyst file from Miriam
catalyts = pd.read_excel('../data/raw/Catalysts.xlsx')

#drop the first two columns (just comments)
catalyts = catalyts.drop(0)
catalyts = catalyts.drop(1)

catalyts = catalyts.rename(columns=catalyts.iloc[0])
catalyts = catalyts.drop(2)
catalyts.tail(5)

Make Miriam excel sheet into easy readable .tsv file

#some catalysts are mixtures of chemicals (then typically the first entry then entries have 'nan' and belong to previous entry)
catalyst_id = 0
current_cat = None

catalysators = {}

for row in catalyts.iterrows():
    
    #Get name, if it is a composite catalyst and the component name
    current_cat = row[1][2]
    is_mix = row[1][1]
    comp_name = row[1][3]
    
    #Get comment
    comment = row[1][7]
    if str(comment) == 'nan':
        comment = 'None'
    comment = comment.strip()


    #Get source
    source = row[1][8]
    if str(source) == 'nan':
        source = 'None'
    source = source.strip()
    
    
    #Get GSH codes
    GSH_codes = [x.strip() for x in row[1][4].split(',') if x != 'X' and x != 'x']
    
    #Get H_Codes (flatten codes combined with Plus)
    H_codes =[x.strip() for x in row[1][5].split(',') if x != 'X' and x != 'x']  
    
    #flatten the whole list
    H_flatten = []
    for H in H_codes:
        H_flatten.append(H.replace(' ',''))
        
            
    
    #Get H_Codes (flatten codes combined with Plus)
    P_codes =[x.strip() for x in row[1][6].split(',') if x != 'X' and x != 'x']  
    #P_flatten = P_codes
    
    P_flatten = []
    for P in P_codes:
        P_flatten.append(P.replace(' ',''))

    

    #if to_annotate == 'nan' it means it is part of a mixture (mixture entries have the first row annotated)
    if str(row[1][0]) != 'nan':
        #if is has a name it is either the start of a mixture (or a single agent catalyst)
        
        #set catalyst and component_ID
        catalyst_id +=1
        compononent_ID = 0
        
        if is_mix == 'mixture':
            catalysators[catalyst_id]= {'Catalyst_Name':current_cat,'Comment':comment,'Is_Mixture':'YES', compononent_ID:{'Component_Name':comp_name, 'GHS':GSH_codes, 'H':H_flatten, 'P':P_flatten, 'Source':source}}
        else:
            catalysators[catalyst_id]= {'Catalyst_Name':current_cat,'Comment':comment,'Is_Mixture':'NO', compononent_ID:{'Component_Name':current_cat, 'GHS':GSH_codes, 'H':H_flatten, 'P':P_flatten, 'Source':source}}
    
    #if here it means further entries of the same catalyst
    else:
        compononent_ID += 1
        catalysators[catalyst_id][compononent_ID] = {'Component_Name':comp_name, 'GHS':GSH_codes, 'H':H_flatten, 'P':P_flatten, 'Source':source}

# CHECK Original List with Miriam/Fabian List (of names of catalysts)
list_original_cats = []

#open list previously compiled in 1_Create_List_All_Solvents.ipynb (containing all reagents)
fp = open('../results/1_Solvent_Catalyst_Lists/List_of_all_reagents_and_catalysts.tsv')
for line in fp:
    cat = line.strip()
    list_original_cats.append(cat)
    
    found = False
    for i in range(1, max(catalysators.keys())+1):
        if cat == catalysators[i]['Catalyst_Name']:
            found = True
            
    if found == False:
        print (cat)
        
# IF not a single NOT FOUND then perfect! This means all catalyts/reagents have been correctly manual annotated.

for i in range(1, max(catalysators.keys())+1):
    if catalysators[i]['Catalyst_Name'] not in list_original_cats:
        print (catalysators[i]['Catalyst_Name'])
# USE to adjust names

#### Make final output file

# MAKE OUTPUT FILE
fp_out = open('../data/processed/Catalysts_Overview.tsv','w')
fp_out.write('Catalyst_ID\tCatalyst_Name\tIs_Mixture\tComponent_ID\tComponent_Name\tGSH_Codes\tH_Codes\tP_Codes\tSource\tComments\n')

#Go through all catalystators
for i in range(1, max(catalysators.keys())+1):
    
    #go through all components
    i2= 0
    while i2 in catalysators[i]:
        
        #print (catalysators[i]['Catalyst_Name'])
        
        fp_out.write(str(i)+'\t'+catalysators[i]['Catalyst_Name']+'\t'+catalysators[i]['Is_Mixture'] +'\t'
                    + str(i2) + '\t' + catalysators[i][i2]['Component_Name'] + '\t' + ';'.join(catalysators[i][i2]['GHS'])
                    + '\t' + ';'.join(catalysators[i][i2]['H']) + '\t' + ';'.join(catalysators[i][i2]['P'])
                    + '\t' + catalysators[i][i2]['Source'] + '\t' + catalysators[i]['Comment'] + '\n')
        
        i2 +=1
    
fp_out.close()

### Read condition file

interesting_conditions = []
fp = open('../results/1_Solvent_Catalyst_Lists/List_of_interesting_conditions.tsv')
for line in fp:
    interesting_conditions.append(line.strip())
print (interesting_conditions)

### Create a complete list of all reactions used in this study

#selected reaction conditions
important_columns = ['Temperature (Reaction Details) [C]','Time (Reaction Details) [h]','Solvent (Reaction Details)','Catalyst','Reagent','Yield (numerical)','Other Conditions','Number of Reaction Steps','Reaction: Links to Reaxys','References']

#open empty dataframe with seleted parameters, i.e., important_columns
all_compounds = pd.DataFrame(columns=important_columns)
all_compounds['Compound_ID'] = ''
all_compounds['Type'] = ''

#go through all compounds, add all found reactions to the initialized empty dataframe
mypath = '../data/raw/compounds/'
onlyfiles = [f for f in listdir(mypath) if isfile(join(mypath, f)) and '~' not in f]
for file in onlyfiles:

    print (file)

    #red the excel file
    compound = pd.read_excel(mypath+file)

    #differentiate between literature and real(Fabians) conditions
    fabian_condition = compound.iloc[len(compound)-3,][important_columns]
    compound = compound.iloc[0:len(compound)-3,][important_columns]
    
    
    #get current compound id
    compound_id = file.split('.')[0].split('_')[2]

    
    #add special row to distinguish the example and fabian conditions
    compound['Type'] = 'Other'
    fabian_condition['Type'] = 'Fabian'
    compound = compound.append(fabian_condition)

    #make Tmp and time to string
    compound['Temperature (Reaction Details) [C]'] = compound['Temperature (Reaction Details) [C]'].astype(str)
    compound['Time (Reaction Details) [h]'] = compound['Time (Reaction Details) [h]'].astype(str)
    
    
    #take higher value for ranges
    compound.loc[compound['Temperature (Reaction Details) [C]'].str.contains(';', na=False), 'Temperature (Reaction Details) [C]'] = compound.loc[compound['Temperature (Reaction Details) [C]'].str.contains(';', na=False)]['Temperature (Reaction Details) [C]'].str.split(';').str[1]
    compound.loc[compound['Temperature (Reaction Details) [C]'].str.contains(' - ', na=False), 'Temperature (Reaction Details) [C]'] = compound.loc[compound['Temperature (Reaction Details) [C]'].str.contains(' - ', na=False)]['Temperature (Reaction Details) [C]'].str.split(' - ').str[1]
    compound.loc[compound['Time (Reaction Details) [h]'].str.contains(';', na=False), 'Time (Reaction Details) [h]'] = compound.loc[compound['Time (Reaction Details) [h]'].str.contains(';', na=False)]['Time (Reaction Details) [h]'].str.split(';').str[1]
    
    #set 'catalyst' to true/false
    
    #compound.Catalyst.loc[(compound['Catalyst'] != False)] = 1
    compound = compound.fillna(value = {'Catalyst':''})
    compound = compound.fillna(value = {'Reagent':''})
    
    
    #replace empty/nan values with the unique 'None' as identifier for missing values
    compound = compound.fillna('None')
    compound = compound.replace({'nan': 'None'})

    
    

    compound['Reagent/Catalyst'] = compound['Catalyst'] + compound['Reagent']
    compound.loc[compound['Reagent'] == '', 'Reagent'] = 'None'
    compound.loc[compound['Catalyst'] == '', 'Catalyst'] = 'None'
    compound.loc[compound['Reagent/Catalyst'] == '', 'Reagent/Catalyst'] = 'None'
    #compound = compound.fillna(value = {'Reagent':''})
    
    
    #add compound ID
    compound['Compound_ID'] = compound_id

    for x in compound.iterrows():
        
        conditions =  x[1][6].split(';')

        for c in conditions:
            c = c.strip()
            if c not in interesting_conditions:
                #print (c)
                compound.iloc[x[0]]['Other Conditions'] = 'None'

    
    #add current compound to all compounds
    all_compounds = all_compounds.append(compound)
    
#save final csv file
all_compounds.to_csv('../data/processed/All_Reactions.csv')
compound.head()
